# swagger.yaml
swagger: "2.0"
info:
  title: Next OMS API
  description: The purpose of this service is to handle order management and facilitate user authentication and authorization.
  version: v1.0.0
  contact:
    name: Rezwanul-Haque
    email: rezwanul.cse@gmail.com
host: localhost:8080
basePath: /api
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Provide the Bearer token with the 'Bearer <token>' format."
paths:
  /v1/orders:
    post:
      tags:
        - Orders
      summary: Create Order
      description: Create a new order.
      operationId: createOrder
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            type: object
            example: {}
      responses:
        "201":
          description: Order created successfully.
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: '#/responses/errorResponse'
      security:
        - BearerAuth: []
  /v1/orders/{orderId}/cancel:
    put:
      tags:
        - Orders
      summary: Cancel Order
      description: Cancel an existing order.
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          required: true
          type: string
          description: The ID of the order to cancel.
      responses:
        "200":
          description: Order canceled successfully.
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: '#/responses/errorResponse'
      security:
        - BearerAuth: []
  /v1/orders/all:
    get:
      tags:
        - Orders
      summary: Get Orders
      description: Retrieve all orders.
      operationId: getOrders
      responses:
        "200":
          description: Orders retrieved successfully.
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: '#/responses/errorResponse'
      security:
        - BearerAuth: []
  /v1/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Login a user.
      operationId: login
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            type: object
            example: { "username": "01901901901@mailinator.com", "password": "321dsa" }
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
  /v1/users/signup:
    post:
      tags:
        - Users
      summary: User Signup
      description: Register a new user.
      operationId: userSignup
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: '#/definitions/UserReq'
      responses:
        "201":
          $ref: '#/responses/UserCreatedResponse'
        "400":
          $ref: '#/responses/errorResponse'
  /v1/logout:
    post:
      tags:
        - Authentication
      summary: Logout
      description: Logout a user.
      operationId: logout
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          description: Bearer token for the authenticated user.
      responses:
        "200":
          description: Logout successful.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
definitions:
  RestErr:
    type: object
    properties:
      error:
        type: string
        example: bad_request
      message:
        type: string
        example: error message
      status:
        type: integer
        format: int64
        example: 400
  UserReq:
    type: object
    properties:
      email:
        type: string
        example: 01901901901@mailinator.com
      first_name:
        type: string
        example: Hello
      last_name:
        type: string
        example: World
      password:
        type: string
        example: 321dsa
      phone:
        type: string
        example: 01901901901
      profile_pic:
        type: string
        example: https://dummyimage.com/300x300/
      user_name:
        type: string
        example: Hello-World
  UserResp:
    type: object
    properties:
      email:
        type: string
      first_login:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
        format: int64
      last_login_at:
        type: string
        format: date-time
      last_name:
        type: string
      phone:
        type: string
      profile_pic:
        type: string
      user_name:
        type: string
responses:
  UserCreatedResponse:
    description: Response after a user is created
    schema:
      $ref: '#/definitions/UserResp'
  errorResponse:
    description: Generic error response
    schema:
      $ref: '#/definitions/RestErr'